rename file and get compiling
To Do
- check that changning how bin edges are passed doesn't seg fault
- add more creation helpers
- check if we are in a subdirectory? share Process handle so we can open
histogram directory for the user upon creation?
